refactor: extract metaconfig utils into dedicated module
